services/gateway/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
httpx>=0.27.0,<1.0.0

# Testing
pytest>=8.0.0,<9.0.0
